Don't serve source files
