+++ sproxy-web.cabal
@@ -75,10 +75,23 @@ test-suite spec
   build-depends:
     base == 4.*,
     hspec,
+    hspec-wai,
     interpolate,
     temporary,
     bytestring,
     configurator,
     hflags,
     directory,
-    filepath
+    filepath,
+    postgresql-simple,
+    http-types,
+    blaze-html,
+    mtl,
+    data-default-class,
+    resource-pool,
+    wai,
+    warp,
+    wai-extra,
+    wai-middleware-static,
+    scotty,
+    text
