+++ test/MainSpec.hs
@@ -1,0 +1,26 @@
+{-# LANGUAGE OverloadedStrings #-}
+
+module MainSpec where
+
+import Test.Hspec
+import Network.Wai
+import Test.Hspec.Wai
+import Web.Scotty.Trans (scottyAppT)
+
+import Config
+import DB
+import Run
+
+app :: IO Application
+app = do
+    pool <- createDBPool "something"
+    sDir <- getStaticDir
+    scottyAppT id id $ sproxyWeb sDir pool
+
+spec :: Spec
+spec = with app $ do
+  describe "sproxyWeb" $ do
+    it "doesn't serve source files" $ do
+      get "/sproxy-web.cabal" `shouldRespondWith` 500
+    it "serve static file" $ do
+      get "/static/loading.gif" `shouldRespondWith` 200
